lock: Use a Guard strategy to support non-Service uses
In order for the `cache` middleware to use linkerd's Lock
implementation, the impl must behave more like a generic
(Service-agnostic) Lock.

This change modifies `Lock` to return a `Guard`; and provides a
`Service` wrapper. This drastically simplifies the lock's state, as
Failure is no longer supported at the lock level.